Remove support for EC aliases in SunPKCS11
Reverting a previous commit (8ddd8d766221c1ac26455512698bcf58d6e24a1c)
that added support for EC aliases in KeyFactory and KeyPairGenerator
for SunPKCS11.

This was previously supported in Java 8 on Z by the IBMPKCS11Impl
provider. However, an openjdk contributer noted that OID is not
commonly used for the KeyFactory and KeyGenerator services.
Instead, the standard "EC" string should be used as an input to
generate EC keys.

Given the upstream stance, no requests for continuing this support,
and the availability of an alternative, this change should be removed.

Signed-off-by: Bob Du <bob.du@ibm.com>